code/requirements.txt
# Core dependencies with pinned versions for reproducibility
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# ML/DL frameworks
xgboost==1.7.6
lightgbm==4.0.0
torch==2.0.1
tensorflow==2.13.0

# Imbalanced learning
imbalanced-learn==0.11.0

# LLM & NLP
openai==0.27.8
anthropic==0.3.10
transformers==4.31.0
sentence-transformers==2.2.2
tiktoken==0.4.0

# Evaluation metrics
rouge-score==0.1.2

# Visualization & diagrams
graphviz==0.20.1
diagrams==0.23.3
kaleido==0.2.1

# Data & utilities
pyyaml==6.0.1
jsonlines==3.1.0
python-dotenv==1.0.0
tqdm==4.65.0
requests==2.31.0

# Privacy & security
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
cryptography==41.0.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-timeout==2.1.0

# Monitoring & logging
structlog==23.1.0
